Add new apt_info.py script
Currently, apt information is extracted via a shell script by way of
`apt-get --just-print dist-upgrade`, and it has a number of issues. For
example, it includes held packages in the metrics, and includes
information about packages that aren't even installed. Replace `apt.sh`
with a new Python-based script that uses the APT API to extract
everything we need, while fixing known issues and remaining mostly
compatible with the old metric format.

Fix #85
Fix #90

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>