NVMe metrics collector based on parallel
The collector based on parallel allows to collect the metrics quicker when
the node has several NVMe drives. Included a Grafana dashboard that visualizes
the metrics and for each panel reports a description coming from the NVMe express
specifications.

Because there are now four files NVMe related everything has been moved to a
common directory.

Signed-off-by: dobbi84 <davide.obbi@e4company.com>