inotify_instance: add commandline arguments for command and user filtering
on systems with many short lived processes, the cardinality of the
rendered metrics can become quite bad.

this change adds commandline arguments to the script to support
filtering of processes based on the process command or user. the
script logic remains backwards compatible as it continues to
emit metrics for all accessible process information if not filters
are provided.

alternatively the rendered metrics can be summed up by user to
reduce the cardinality even further.

Signed-off-by: Gordon Bleux <33967640+UiP9AV6Y@users.noreply.github.com>