smartmon.sh: locate smartctl tool on non-Linux systems (#240)
Allow execution of `smartctl` even if it is installed in an unexpected
(for Linux) path.

Some OS, (e.g.: FreeBSD) do not install `smartctl` under the `usr/sbin/`
path, but instead use another path like `/usr/local/sbin/`.
This commit makes use of the `which` command to locate the `smartctl` in
the `PATH` and execute from there. As such it does not make any
assumption where `smartctl` is installed, improving portability.

---------

Signed-off-by: cristian-caloghera <33933948+cristian-caloghera@users.noreply.github.com>